# Add support for using VAES instructions for NI parallel operations.