--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,6 @@
 target/
 **/Cargo.lock
 **/target/
+**/pin.log
+**/pin-log.txt
+**/pin-tool-log.txt