--- a/aes/Cargo.toml
+++ b/aes/Cargo.toml
@@ -18,7 +18,7 @@ cipher = "=0.5.0-pre.2"
 zeroize = { version = "1.5.6", optional = true, default_features = false, features = ["aarch64"] }
 
 [target.'cfg(any(target_arch = "aarch64", target_arch = "x86_64", target_arch = "x86"))'.dependencies]
-cpufeatures = "0.2"
+cpufeatures = "0.2.12"
 
 [dev-dependencies]
 cipher = { version = "=0.5.0-pre.2", features = ["dev"] }