--- a/aes/src/x86/vaes256.rs
+++ b/aes/src/x86/vaes256.rs
@@ -1,0 +1,3 @@
+pub(super) mod aes128;
+pub(super) mod aes192;
+pub(super) mod aes256;