… .github/workflows/aes.yml
… Cargo.lock
… aes/src/ni.rs
… aes/src/x86.rs
… aes/src/x86/ni/aes192.rs
… aes/src/x86/ni/aes256.rs
… aes/src/x86/ni/hazmat.rs
… aes/src/x86/ni/test_expand.rs
… aes/src/x86/ni/utils.rs
… aes/src/x86/vaes256/aes128.rs
… aes/src/x86/vaes256/aes192.rs
… aes/src/x86/vaes256/aes256.rs
… aes/src/x86/vaes512/aes128.rs
… aes/src/x86/vaes512/aes192.rs
… aes/src/x86/vaes512/aes256.rs