@silvanshade 2024-01-09
> @silvanshade I'd definitely recommend trying to get `cpufeatures` working. The latest v0.2.12 release just added support for detecting VAES.

Yeah, I had created the PR that added that, since I was originally going to try and use `cpufeatures` for this.

> You'd need to add detection for VAES, and a branch to use it if available.

I can create a branch like this. That's not the main difficulty, as I understand it.

Rather, the `VAES` features aren't usable without also enabling `stdsimd` and `avx512_target_feature`, which I added to the top of `lib.rs`.

I could change it to where there is a `--cfg vaes` rather than the cargo feature, so it works more like the `aes_armv8` or something, but there still needs to be something to conditionally enable those features, otherwise `aes` won't even be able to compile on `stable` anymore.

I think maybe the only way to add support for this without some sort of feature gating would be to locally define stable versions of the `VAES` intrinsics like what was done for the ARM backend.

What do you think?