@silvanshade 2024-01-09
> > I think maybe the only way to add support for this without some sort of feature gating would be to locally define stable versions of the VAES intrinsics like what was done for the ARM backend.
> 
> That sounds great!

So it is possible to write local versions of these intrinsics, e.g.,

```rust
#[inline]
#[target_feature(enable = "avx512f")]
pub(super) unsafe fn pf_mm512_aesdec_epi128(data: __m512i, round_key: __m512i) -> __m512i {
    let result: __m512i;
    asm!(
        "vaesdec {result}, {data}, {round_key}",
        data = in(zmm_reg) data,
        round_key = in(zmm_reg) round_key,
        result = out(zmm_reg) result,
        options(pure, nomem, nostack, preserves_flags)
    );
    result
}
```

But what I didn't realize is that it's still necessary to have the `#[target_feature(enable = "avx512f")]`, which in turn requires `#![feature(avx512_target_feature)]` (which is `unstable`) in order to use the AVX registers. So I think at minimum we would probably need to add a `"nightly"` feature or something to gate these on and can't quite get away with just autodetection and local versions of the intrinsics.

If that seems reasonable, I will add that feature (`"nightly"`, or `"unstable"`, or whatever else you'd prefer to call it) and then implement the autodetection for `vaes`.