@silvanshade 2024-01-09
> I think it will be better to write a separate implementation in the `vaes` module instead of piggybacking on the `ni` module. It's also probably worth to increase number of blocks processed in parallel for the VAES backend. Right now, you call only two `aesdec`/`aesenc` functions per round, thus potentially loosing on additional ILP-based throughput (the instructions have latency of 3 cycles and throughput of 1 cycle). Additionally, with AVX-512 you have 32 ZMM registers, so you have less register pressure.

Initially I was planning to do that but the reason I opted not to is because, for the single block case, we still basically want to fall back to the `NI` implementation, don't we? Maybe for maintenance or structural reasons it would be cleaner to just duplicate that code though?

I did try increasing the parallel blocks to 32 (calling 8 of the respective instructions) but didn't notice a performance difference in the benchmarks here, although I only have the one system to test one. But I agree it probably makes sense in general, and especially for a separate backend.