@newpavlov 2024-01-09
>for the single block case, we still basically want to fall back to the NI implementation, don't we? Maybe for maintenance or structural reasons it would be cleaner to just duplicate that code though?

Yes, also key expansion code will be the same. But I think that the parallel processing function definitely should live in the `vaes` module. It will allow us to change number of blocks processed in parallel, will remove the unnecessary key broadcasts and casting between `__m512i` and `__m256i`.

So I think we should define separate backends (i.e. structs which implement the `BlockBackend` trait) which will use functions from the `ni` module for single block processing and key expansion.